feat: move info content to sidebar
As there is no way to move content to Google Maps InfoWindow without
using DOM manipulation, I opted to integrate it into the sidebar.
The toggle button is reused to navigate between the list of stations and
station details.
There was also a bug in the timepicker Knockout bindings that required
some modification.